+++ CHANGELOG.md
@@ -3,3 +3,11 @@
 ## 0.0.0.0 -- 2025-02-05
 
 * Initial version: project skeleton for recalculation engine + test suite
+
+## 0.1.0.0 -- 2025-02-05
+
+* First implementation: minimal recalculation engine for generic language
+  implementation based on "Build systems à la carte". Includes a simple
+  example language as part of the test-suite.
+
+  Add [`README.md`](./README.md) and document the project organisation a bit.